+++ js/data/data.js
@@ -1,7 +1,3 @@
-import {getRandomArrayElement} from './utils';
-import {createIdGenerator} from './utils';
-import {getRandomInteger} from './utils';
-
 const COMMENTS = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -70,25 +66,3 @@ const DESCRIPTIONS = [
   'Скрытые детали',
 ];
 const PHOTO_COUNT = 25;
-const generatePhotoId = createIdGenerator();
-const generateUrl = createIdGenerator();
-const generateCommentId = createIdGenerator();
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: `${getRandomArrayElement(COMMENTS)}`,
-  name: `${getRandomArrayElement(NAMES)}`,
-});
-
-const createPhoto = () => ({
-  id: generatePhotoId(),
-  url: `photos/${generateUrl()}.jpg`,
-  description: `${getRandomArrayElement(DESCRIPTIONS)}`,
-  likes: getRandomInteger(15, 200),
-  comments: Array.from({length: getRandomInteger(0, 30)}, createComment),
-});
-
-const createPhotos = Array.from({length: PHOTO_COUNT}, createPhoto);
-
-export {createPhotos};
